Split arithmetic ops from routines/math.h
