+++ chainerx_cc/chainerx/cuda/cuda_device/misc.cu
@@ -13,6 +13,7 @@
 #include "chainerx/cuda/numeric.cuh"
 #include "chainerx/device.h"
 #include "chainerx/dtype.h"
+#include "chainerx/kernels/arithmetic.h"
 #include "chainerx/kernels/math.h"
 #include "chainerx/kernels/misc.h"
 #include "chainerx/routines/type_util.h"
