+++ chainerx_cc/chainerx/routines/indexing.cc
@@ -17,9 +17,11 @@
 #include "chainerx/constant.h"
 #include "chainerx/dtype.h"
 #include "chainerx/graph.h"
+#include "chainerx/kernels/arithmetic.h"
 #include "chainerx/kernels/indexing.h"
 #include "chainerx/kernels/math.h"
 #include "chainerx/macro.h"
+#include "chainerx/routines/arithmetic.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/routines/type_util.h"
