+++ chainerx_cc/chainerx/routines/normalization.cc
@@ -19,6 +19,7 @@
 #include "chainerx/graph.h"
 #include "chainerx/kernels/normalization.h"
 #include "chainerx/macro.h"
+#include "chainerx/routines/arithmetic.h"
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/routines/misc.h"
