+++ chainerx_cc/chainerx/routines/pooling.cc
@@ -19,6 +19,7 @@
 #include "chainerx/error.h"
 #include "chainerx/graph.h"
 #include "chainerx/kernels/pooling.h"
+#include "chainerx/routines/arithmetic.h"
 #include "chainerx/routines/math.h"
 #include "chainerx/routines/routines_util.h"
 #include "chainerx/stack_vector.h"
